backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pandas==2.2.0
python-multipart==0.0.6
openpyxl==3.1.2
pydantic-settings==2.1.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
